py-scripts/requirements.txt
altgraph==0.17.5
Appium-Python-Client==5.2.4
attrs==25.4.0
certifi==2026.1.4
customtkinter==5.2.2
exceptiongroup==1.3.1
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
macholib==1.16.4
outcome==1.3.0.post0
packaging==25.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2025.11
PySocks==1.7.1
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.2.0
zipp==3.23.0
